Moved all saving functionality into useCharacter* hooks, validator into helper functions to return to
